hello_world/requirements.txt
requests
aiohttp
pydantic
openai
python-ulid
dynamodb-json